requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
Pillow==10.3.0
pdf2image==1.17.0
google-genai # For the Multimodal LLM API